Split riscv example into 2
